Extract consumer to separate package
